refactor(datacollect): extract V1 adapter support, drop built-in mock-sync-server
